+++ AWTEditor/src/AWT/UI/AWTToggleLayersMenu.java
@@ -41,7 +41,7 @@ public class AWTToggleLayersMenu extends AWTDropdownMenu {
 		list.setButtons(menuButtons);
 		list.setPostition(new Point(142,2 + optionsButton.getHeight()));
 		list.setButtonOffset(2);
-		list.setButtonDimensions(optionsButton.getWidth() - 4, optionsButton.getHeight() - 4);
+		list.setButtonDimensions((int)optionsButton.getWidth() - 4, (int)optionsButton.getHeight() - 4);
 		
 		setRoot(optionsButton);
 		setMenu(list);
